+++ src/metier/entite/question/qcm/QCM.java
@@ -18,7 +18,7 @@ public class QCM extends Question
 	/*--------------*/
 	/* Constructeur */
 	/*--------------*/
-	public QCM(int id, Ressource ressource, Notion notion, Difficulte difficulte, int temps, int note,boolean unique)
+	public QCM(int id, Ressource ressource, Notion notion, Difficulte difficulte, int temps, double note, boolean unique)
 	{
 		super(id, ressource, notion, difficulte, temps, note);
 		this.unique = unique;
